xgets: rewrite in terms of C++ strings
Update `xgets` to return a std::string, instead of a malloc'ed
character buffer.  Also update the logic to, hopefully, make it
somewhat clearer.  Call sites were updated accordingly.

Also, the logic for detecting aborted input (via the intr
character) was broken; ^C always presented as EOF.  Changing
this also afforded the opportunity to change the `ignoreeof`
logic to just use a flag, instead of `stty`.

So far, this seems to work, but testing it in a principled way
is challenging because it is deeply complected with so much
global state (in particular, the way the stdin stack works).

This change also rewrites `ngets` in terms of `xgets`, and
removes the `oldngets` function.

Signed-off-by: Dan Cross <cross@m-net.arbornet.org>